Aligned NewBlockHandler with Maru, adjusted the existing protocols
