+++ app/src/main/kotlin/maru/app/MaruApp.kt
@@ -21,7 +21,6 @@ import kotlin.time.Duration.Companion.seconds
 import maru.config.MaruConfig
 import maru.consensus.ForksSchedule
 import maru.consensus.MetadataOnlyHandlerAdapter
-import maru.consensus.NewBlockHandler
 import maru.consensus.NewBlockHandlerMultiplexer
 import maru.consensus.NextBlockTimestampProviderImpl
 import maru.consensus.OmniProtocolFactory
@@ -39,7 +38,6 @@ class MaruApp(
   config: MaruConfig,
   beaconGenesisConfig: ForksSchedule,
   clock: Clock = Clock.systemUTC(),
-  newBlockHandlers: Map<String, NewBlockHandler> = emptyMap(),
 ) {
   private val log: Logger = LogManager.getLogger(this::class.java)
 
@@ -58,7 +56,7 @@ class MaruApp(
         .toString(),
     )
 
-  private val newBlockHandlerMultiplexer = NewBlockHandlerMultiplexer(newBlockHandlers)
+  private val newBlockHandlerMultiplexer = NewBlockHandlerMultiplexer(emptyMap())
 
   private val metadataProvider = Web3jMetadataProvider(ethereumJsonRpcClient.eth1Web3j)
