+++ consensus/src/main/kotlin/maru/consensus/delegated/Mapper.kt
@@ -15,53 +15,141 @@
  */
 package maru.consensus.delegated
 
+import fromHexToByteArray
+import java.math.BigInteger
+import maru.core.BeaconBlock
+import maru.core.BeaconBlockBody
+import maru.core.BeaconBlockHeader
+import maru.core.ExecutionPayload
+import maru.core.HashUtil
+import maru.core.Validator
+import maru.serialization.rlp.KeccakHasher
+import maru.serialization.rlp.RLPSerializers
 import org.apache.tuweni.bytes.Bytes
-import org.apache.tuweni.bytes.Bytes32
+import org.bouncycastle.math.ec.custom.sec.SecP256K1Curve
+import org.hyperledger.besu.crypto.SECPSignature
+import org.hyperledger.besu.datatypes.AccessListEntry
 import org.hyperledger.besu.datatypes.Address
-import org.hyperledger.besu.datatypes.BlobGas
-import org.hyperledger.besu.datatypes.Hash
 import org.hyperledger.besu.datatypes.Wei
-import org.hyperledger.besu.ethereum.core.Block
-import org.hyperledger.besu.ethereum.core.BlockBody
-import org.hyperledger.besu.ethereum.core.BlockHeader
-import org.hyperledger.besu.ethereum.core.Difficulty
-import org.hyperledger.besu.ethereum.mainnet.BodyValidation
-import org.hyperledger.besu.ethereum.mainnet.MainnetBlockHeaderFunctions
-import org.hyperledger.besu.evm.log.LogsBloomFilter
+import org.hyperledger.besu.ethereum.core.Transaction
+import org.hyperledger.besu.ethereum.core.encoding.EncodingContext
+import org.hyperledger.besu.ethereum.core.encoding.TransactionEncoder
 import org.web3j.protocol.core.methods.response.EthBlock
+import org.web3j.protocol.core.methods.response.EthBlock.TransactionObject
 
 object Mapper {
-  private val blockHeaderFunctions = MainnetBlockHeaderFunctions()
+  private val hasher = HashUtil.headerHash(RLPSerializers.BeaconBlockHeaderSerializer, KeccakHasher)
 
-  fun mapWeb3jBlockToBesuBlock(block: EthBlock.Block): Block {
-    val blockHeader =
-      BlockHeader(
-        /* parentHash = */ Hash.wrap(Bytes32.fromHexString(block.parentHash)),
-        /* ommersHash = */ Hash.EMPTY_LIST_HASH,
-        /* coinbase = */ Address.fromHexString(block.miner),
-        /* stateRoot = */ Hash.wrap(Bytes32.fromHexString(block.stateRoot)),
-        /* transactionsRoot = */ Hash.wrap(Bytes32.fromHexString(block.transactionsRoot)),
-        /* receiptsRoot = */ Hash.wrap(Bytes32.fromHexString(block.receiptsRoot)),
-        /* logsBloom = */ LogsBloomFilter(Bytes.fromHexString(block.logsBloom)),
-        /* difficulty = */ Difficulty.ZERO,
-        /* number = */ block.number.toLong(),
-        /* gasLimit = */ block.gasLimit.toLong(),
-        /* gasUsed = */ block.gasUsed.toLong(),
-        /* timestamp = */ block.timestamp.toLong(),
-        /* extraData = */ Bytes.fromHexString(block.extraData),
-        /* baseFee = */ Wei.of(block.baseFeePerGas.toLong()),
-        /* mixHashOrPrevRandao = */ Bytes32.fromHexString(block.mixHash),
-        /* nonce = */ 0,
-        /* withdrawalsRoot = */ Hash.EMPTY_TRIE_HASH,
-        /* blobGasUsed = */ 0,
-        /* excessBlobGas = */ BlobGas.ZERO,
-        /* parentBeaconBlockRoot = TODO: use an actual beacon block root */ Bytes32.ZERO,
-        /* requestsHash = */ BodyValidation.requestsHash(listOf()),
-        /* blockHeaderFunctions = */ blockHeaderFunctions,
+  fun mapWeb3jBlockToBeaconBlock(block: EthBlock.Block): BeaconBlock {
+    val transactions =
+      block.transactions.map {
+        val transaction = it.get() as TransactionObject
+        transaction.toBytes()
+      }
+    val executionPayload =
+      ExecutionPayload(
+        parentHash = block.parentHash.fromHexToByteArray(),
+        feeRecipient = block.miner.fromHexToByteArray(),
+        stateRoot = block.stateRoot.fromHexToByteArray(),
+        receiptsRoot = block.receiptsRoot.fromHexToByteArray(),
+        logsBloom = block.logsBloom.fromHexToByteArray(),
+        prevRandao = block.mixHash.fromHexToByteArray(),
+        blockNumber = block.number.toLong().toULong(),
+        gasLimit = block.gasLimit.toLong().toULong(),
+        gasUsed = block.gasUsed.toLong().toULong(),
+        timestamp = block.timestamp.toLong().toULong(),
+        extraData = block.extraData.fromHexToByteArray(),
+        baseFeePerGas = block.baseFeePerGas,
+        blockHash = block.hash.fromHexToByteArray(),
+        transactions = transactions,
+      )
+    val beaconBlockBody = BeaconBlockBody(prevCommitSeals = emptyList(), executionPayload = executionPayload)
+
+    val beaconBlockHeader =
+      BeaconBlockHeader(
+        number = 0u,
+        round = 0u,
+        timestamp = block.timestamp.toLong().toULong(),
+        proposer = Validator(block.miner.fromHexToByteArray()),
+        parentRoot = ByteArray(32),
+        stateRoot = ByteArray(32),
+        bodyRoot = ByteArray(32),
+        headerHashFunction = hasher,
+      )
+    return BeaconBlock(beaconBlockHeader, beaconBlockBody)
+  }
+
+  private fun recIdFromV(v: BigInteger): Pair<Byte, BigInteger?> {
+    val recId: Byte
+    var chainId: BigInteger? = null
+    if (
+      v == Transaction.REPLAY_UNPROTECTED_V_BASE ||
+      v == Transaction.REPLAY_UNPROTECTED_V_BASE_PLUS_1
+    ) {
+      recId = v.subtract(Transaction.REPLAY_UNPROTECTED_V_BASE).byteValueExact()
+    } else if (v > Transaction.REPLAY_PROTECTED_V_MIN) {
+      chainId = v.subtract(Transaction.REPLAY_PROTECTED_V_BASE).divide(Transaction.TWO)
+      recId =
+        v
+          .subtract(Transaction.TWO.multiply(chainId).add(Transaction.REPLAY_PROTECTED_V_BASE))
+          .byteValueExact()
+    } else {
+      throw RuntimeException("An unsupported encoded `v` value of $v was found")
+    }
+    return Pair(recId, chainId)
+  }
+
+  // TODO: Test
+  private fun TransactionObject.toBytes(): ByteArray {
+    val isFrontier = this.type == "0x0"
+    val (recId, chainId) =
+      if (isFrontier) {
+        recIdFromV(this.v.toBigInteger())
+      } else {
+        Pair(this.v.toByte(), BigInteger.valueOf(this.chainId))
+      }
+    val signature =
+      SECPSignature.create(
+        BigInteger(this.r.removePrefix("0x"), 16),
+        BigInteger(this.s.removePrefix("0x"), 16),
+        recId,
+        SecP256K1Curve().order,
       )
 
-    // Transactions are omitted
-    val blockBody = BlockBody(listOf(), listOf())
-    return Block(blockHeader, blockBody)
+    val transaction =
+      Transaction
+        .builder()
+        .nonce(this.nonce.toLong())
+        .also { builder ->
+          if (isFrontier || this.type == "0x1") {
+            builder.gasPrice(Wei.of(this.gasPrice))
+          } else {
+            builder.maxPriorityFeePerGas(Wei.of(this.maxPriorityFeePerGas))
+            builder.maxFeePerGas(Wei.of(this.maxFeePerGas))
+          }
+        }.gasLimit(this.gas.toLong())
+        .to(Address.fromHexString(this.to))
+        .value(Wei.of(this.value))
+        .signature(signature)
+        .payload(Bytes.fromHexString(this.input))
+        .also { builder ->
+          this.accessList?.also { accessList ->
+            builder.accessList(
+              accessList.map { entry ->
+                AccessListEntry.createAccessListEntry(
+                  Address.fromHexString(entry.address),
+                  entry.storageKeys,
+                )
+              },
+            )
+          }
+        }.sender(Address.fromHexString(this.from))
+        .apply {
+          if (chainId != null) {
+            chainId(chainId)
+          }
+        }.build()
+
+    return TransactionEncoder.encodeOpaqueBytes(transaction, EncodingContext.BLOCK_BODY).toArray()
   }
 }
